+++ dimcli/core/utils.py
@@ -272,19 +272,6 @@ def open_multi_platform(fpath):
             print("Couldnt find suitable opener for %s" % fpath)
 
 
-def get_dimensions_url(obj_id, obj_type):
-    """
-    Generate a valid Dimensions URL for one of the available sources.
-    obj_id: the dimensions ID of the object
-    obj_type: one of 'publications', 'grants', 'patents', 'policy_documents', 'clinical_trials', 'researchers'
-    """
-    if obj_type not in G.sources():
-        print("Valid sources are: ", " ".join([x for x in G.sources()]))
-        return 
-    url = G.url_for_source(obj_type)
-    if url:
-        return url + obj_id
-
 
 def init_config_folder(user_dir, user_config_file):
     """
@@ -413,7 +400,7 @@ def preview_results(jsondata, maxitems=10):
                 counter += 1
                 if counter <= maxitems:
                     try:  # title and url/id if object has them
-                        url = get_dimensions_url(row['id'], key) or row['id']
+                        url = dimensions_url(row['id'], key) or row['id']
                         if 'title' in row.keys():
                             name_or_title = row['title'].strip()
                         else:
@@ -464,3 +451,39 @@ def print_json_full(jsondata):
 
 
 
+
+def dimensions_url(obj_id, obj_type):
+    """
+    Generate a valid Dimensions URL for one of the available sources.
+    obj_id: the dimensions ID of the object
+    obj_type: one of 'publications', 'grants', 'patents', 'policy_documents', 'clinical_trials', 'researchers'
+    """
+    if obj_type not in G.sources():
+        print("Valid sources are: ", " ".join([x for x in G.sources()]))
+        return 
+    url = G.url_for_source(obj_type)
+    if url:
+        return url + obj_id
+
+
+def dimensions_url(obj_id, obj_type):
+    """
+    Generate a valid Dimensions URL for one of the available sources.
+    obj_id: the dimensions ID of the object
+    obj_type: one of 'publications', 'grants', 'patents', 'policy_documents', 'clinical_trials', 'researchers'
+    """
+    if obj_type not in G.sources():
+        print("Valid sources are: ", " ".join([x for x in G.sources()]))
+        return 
+    url = G.url_for_source(obj_type)
+    if url:
+        return url + obj_id
+
+
+def google_url(stringa):
+    """
+    Generate a valid google search URL from a string (URL quoting is applied)
+    """
+    from urllib.parse import quote   
+    s = quote(stringa)    
+    return f"https://www.google.com/search?q={s}"