+++ dimcli/main_cli.py
@@ -76,7 +76,7 @@ def main_cli(ctx, instance_name=None, init=False, config=False, versioncheck=Fal
         # try online version check
         test = is_dimcli_outdated()
         if test:
-            click.secho("====\nWarning: there is a newer version of Dimcli available at https://pypi.org/project/dimcli/.\n====", bold=True)
+            click.secho("====\nHeads up: there is a newer version of Dimcli available at https://pypi.org/project/dimcli/.\n====", bold=True)
         # launch REPL
         repl.run(instance_name)
     else: