+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## unreleased
+
+* `dsl_escape` util
+
+
 
 ## v 0.5.8